fixtodo: move node status condition code to kubelet_node_conditions.go to be centralized
