+++ pkg/kubelet/BUILD
@@ -14,6 +14,7 @@ go_library(
         "kubelet.go",
         "kubelet_getters.go",
         "kubelet_network.go",
+        "kubelet_node_conditions.go",
         "kubelet_node_status.go",
         "kubelet_pods.go",
         "kubelet_resources.go",
@@ -148,6 +149,7 @@ go_test(
         "active_deadline_test.go",
         "kubelet_getters_test.go",
         "kubelet_network_test.go",
+        "kubelet_node_conditions_test.go",
         "kubelet_node_status_test.go",
         "kubelet_pods_test.go",
         "kubelet_resources_test.go",
